[rename] button js file.
